address comments
